+++ TractorMod/Framework/TractorSoundType.cs
@@ -1,0 +1,15 @@
+namespace Pathoschild.Stardew.TractorMod.Framework
+{
+    /// <summary>The type of sound effects to play while on the tractor.</summary>
+    internal enum TractorSoundType
+    {
+        /// <summary>Disable all sound effects.</summary>
+        None,
+
+        /// <summary>Play the same sound effects as a normal horse.</summary>
+        Horse,
+
+        /// <summary>Play custom tractor sound effects.</summary>
+        Tractor
+    }
+}